refactor: Move all source code into 'src/'
The index.ts now lives among the rest of the source code in 'src/',
instead of weirdly being placed among tool configs.

This patch also contains additional modifications such as removal of the
second .eslintrc.yml in 'test/' in favor of an override, changing the
globs in TSConfig from 'exclude' to 'include', and setting a text
reporter for c8 in addition to the lcov reporter.
