+++ .eslintrc.yml
@@ -4,3 +4,12 @@ extends:
 
 parserOptions:
   project: './tsconfig.eslint.json'
+
+overrides:
+  - files: 'test/**/*'
+    env:
+      mocha: true
+    rules:
+      no-unused-expressions: off
+      '@typescript-eslint/no-unused-expressions': off
+      '@typescript-eslint/no-floating-promises': off
