--- index.ts
@@ -1,3 +1,0 @@
-export { Adapter } from './lib/adapter'
-export { MemoryAdapter } from './lib/memory-adapter'
-export { DirectoryAdapter } from './lib/directory-adapter'
