+++ package.json
@@ -13,7 +13,7 @@
     "test": "mocha --require ts-node/register --recursive \"test/**/*.test.*\"",
     "lint": "tsc --noEmit && eslint --ignore-path .gitignore .",
     "lint-fix": "tsc --noEmit && eslint --fix --ignore-path .gitignore .",
-    "coverage": "c8 --reporter=lcov --all --include \"lib/**/*.ts\" npm test",
+    "coverage": "c8 --all --src=src --reporter=text --reporter=lcov npm test",
     "prepare": "npm run build"
   },
   "repository": {
