+++ src/directory-adapter.ts
