+++ src/index.ts
@@ -1,0 +1,3 @@
+export { Adapter } from './adapter'
+export { MemoryAdapter } from './memory-adapter'
+export { DirectoryAdapter } from './directory-adapter'
