+++ src/memory-adapter.ts
